:hammer: File list display, pictures support direct preview
