+++ composeApp/src/commonMain/kotlin/com/clipevery/ui/resource/ClipResourceLoader.kt
@@ -1,0 +1,8 @@
+package com.clipevery.ui.resource
+
+import androidx.compose.ui.ExperimentalComposeUiApi
+import androidx.compose.ui.res.ResourceLoader
+
+@OptIn(ExperimentalComposeUiApi::class)
+interface ClipResourceLoader: ResourceLoader {
+}
