+++ composeApp/src/desktopMain/kotlin/com/clipevery/ui/resource/DesktopAbsoluteClipResourceLoader.kt
@@ -1,0 +1,11 @@
+package com.clipevery.ui.resource
+
+import java.io.File
+import java.io.InputStream
+
+object DesktopAbsoluteClipResourceLoader: ClipResourceLoader {
+
+    override fun load(resourcePath: String): InputStream {
+        return File(resourcePath).inputStream()
+    }
+}
